Develop

+++ Studdype_Final/src/main/java/com/studdype/test/model/biz/file/photoFileBizImpl.java
@@ -4,7 +4,9 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import com.studdype.test.common.FileHandler;
 import com.studdype.test.model.dao.board.photo.PhotoFileDao;
 import com.studdype.test.model.dto.board.FileDto;
 
@@ -13,6 +15,7 @@ public class photoFileBizImpl implements photoFileBiz{
 
 	@Autowired
 	private PhotoFileDao photoFileDao;
+	private FileHandler fileHandler = new FileHandler();
 
 	@Override
 	public List<FileDto> attachImageList(int b_no) {
@@ -23,10 +26,31 @@ public class photoFileBizImpl implements photoFileBiz{
 	public List<FileDto> galleryDetailFile(int b_no) {
 		return photoFileDao.galleryDetailFile(b_no);
 	}
-
+	
+	@Transactional
 	@Override
 	public int deleteGalleryFile(int f_no) {
-		return photoFileDao.deleteGalleryFile(f_no);
+		int res = 0;
+		
+		FileDto dto = photoFileDao.selectFile(f_no); //파일 가져오기
+		
+		res = fileHandler.deleteFile(dto);
+		
+		if( res > 0 ) {
+			res = photoFileDao.deleteGalleryFile(f_no);
+			
+		}
+		
+		if(dto == null || res < 1) {
+			throw new RuntimeException("글 가져오기 에러");
+		}		
+		return res;
+	}
+	
+
+	@Override
+	public FileDto getFileByFno(int f_no) {
+		return photoFileDao.selectFile(f_no);
 	}
